feat: restrict delete booking to the issuer only

+++ src/main/java/com/example/eventmanagement/service/impl/BookingServiceImpl.java
@@ -98,11 +98,14 @@ public class BookingServiceImpl implements BookingService {
 
     @Override
     public void deleteBooking(Long bookingId) {
+        String email = ((User) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUsername();
         Optional<BookingEntity> booking = bookingRepository.findById(bookingId);
         if (booking.isPresent()) {
+            if (!booking.get().getUser().getEmail().equals(email)) {
+                throw new IllegalArgumentException("Unauthorized access");
+            }
             bookingRepository.delete(booking.get());
-        } else {
-            throw new IllegalArgumentException("Booking not found");
         }
+        throw new IllegalArgumentException("Booking not found");
     }
 }
